+++ maxcompute-20220104/core/Models/DeleteProjectResponseBody.cs
@@ -1,0 +1,34 @@
+// This file is auto-generated, don't edit it. Thanks.
+
+using System;
+using System.Collections.Generic;
+using System.IO;
+
+using Tea;
+
+namespace AlibabaCloud.SDK.MaxCompute20220104.Models
+{
+    public class DeleteProjectResponseBody : TeaModel {
+        /// <summary>
+        /// <para>The returned data.</para>
+        /// 
+        /// <b>Example:</b>
+        /// <para>success</para>
+        /// </summary>
+        [NameInMap("data")]
+        [Validation(Required=false)]
+        public string Data { get; set; }
+
+        /// <summary>
+        /// <para>The ID of the request.</para>
+        /// 
+        /// <b>Example:</b>
+        /// <para>0be3e0aa16667684362147582e038f</para>
+        /// </summary>
+        [NameInMap("requestId")]
+        [Validation(Required=false)]
+        public string RequestId { get; set; }
+
+    }
+
+}