+++ maxcompute-20220104/core/Models/QueryQuotaMetricRequest.cs
@@ -1,0 +1,66 @@
+// This file is auto-generated, don't edit it. Thanks.
+
+using System;
+using System.Collections.Generic;
+using System.IO;
+
+using Tea;
+
+namespace AlibabaCloud.SDK.MaxCompute20220104.Models
+{
+    public class QueryQuotaMetricRequest : TeaModel {
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>60</para>
+        /// </summary>
+        [NameInMap("interval")]
+        [Validation(Required=false)]
+        public long? Interval { get; set; }
+
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>os_sns_p</para>
+        /// </summary>
+        [NameInMap("nickname")]
+        [Validation(Required=false)]
+        public string Nickname { get; set; }
+
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>os_sns</para>
+        /// </summary>
+        [NameInMap("subQuotaNickname")]
+        [Validation(Required=false)]
+        public string SubQuotaNickname { get; set; }
+
+        /// <summary>
+        /// <para>This parameter is required.</para>
+        /// 
+        /// <b>Example:</b>
+        /// <para>1735536322</para>
+        /// </summary>
+        [NameInMap("endTime")]
+        [Validation(Required=false)]
+        public long? EndTime { get; set; }
+
+        /// <summary>
+        /// <para>This parameter is required.</para>
+        /// 
+        /// <b>Example:</b>
+        /// <para>1735534322</para>
+        /// </summary>
+        [NameInMap("startTime")]
+        [Validation(Required=false)]
+        public long? StartTime { get; set; }
+
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>max</para>
+        /// </summary>
+        [NameInMap("strategy")]
+        [Validation(Required=false)]
+        public string Strategy { get; set; }
+
+    }
+
+}