+++ maxcompute-20220104/core/Models/QueryQuotaMetricResponseBody.cs
@@ -1,0 +1,86 @@
+// This file is auto-generated, don't edit it. Thanks.
+
+using System;
+using System.Collections.Generic;
+using System.IO;
+
+using Tea;
+
+namespace AlibabaCloud.SDK.MaxCompute20220104.Models
+{
+    public class QueryQuotaMetricResponseBody : TeaModel {
+        [NameInMap("data")]
+        [Validation(Required=false)]
+        public QueryQuotaMetricResponseBodyData Data { get; set; }
+        public class QueryQuotaMetricResponseBodyData : TeaModel {
+            [NameInMap("category")]
+            [Validation(Required=false)]
+            public string Category { get; set; }
+
+            [NameInMap("metrics")]
+            [Validation(Required=false)]
+            public List<QueryQuotaMetricResponseBodyDataMetrics> Metrics { get; set; }
+            public class QueryQuotaMetricResponseBodyDataMetrics : TeaModel {
+                [NameInMap("metric")]
+                [Validation(Required=false)]
+                public Dictionary<string, string> Metric { get; set; }
+
+                [NameInMap("values")]
+                [Validation(Required=false)]
+                public List<List<double?>> Values { get; set; }
+
+            }
+
+            /// <summary>
+            /// <b>Example:</b>
+            /// <para>cpu</para>
+            /// </summary>
+            [NameInMap("name")]
+            [Validation(Required=false)]
+            public string Name { get; set; }
+
+            /// <summary>
+            /// <b>Example:</b>
+            /// <para>60</para>
+            /// </summary>
+            [NameInMap("period")]
+            [Validation(Required=false)]
+            public long? Period { get; set; }
+
+        }
+
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>success</para>
+        /// </summary>
+        [NameInMap("errorCode")]
+        [Validation(Required=false)]
+        public string ErrorCode { get; set; }
+
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>plan \&quot;***\&quot; does not exist</para>
+        /// </summary>
+        [NameInMap("errorMsg")]
+        [Validation(Required=false)]
+        public string ErrorMsg { get; set; }
+
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>200</para>
+        /// </summary>
+        [NameInMap("httpCode")]
+        [Validation(Required=false)]
+        public int? HttpCode { get; set; }
+
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>0b87b7b316643495896551555e855b</para>
+        /// </summary>
+        [NameInMap("requestId")]
+        [Validation(Required=false)]
+        public string RequestId { get; set; }
+
+    }
+
+}