+++ maxcompute-20220104/core/Models/UpdateProjectModelTierResponseBody.cs
@@ -1,0 +1,57 @@
+// This file is auto-generated, don't edit it. Thanks.
+
+using System;
+using System.Collections.Generic;
+using System.IO;
+
+using Tea;
+
+namespace AlibabaCloud.SDK.MaxCompute20220104.Models
+{
+    public class UpdateProjectModelTierResponseBody : TeaModel {
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>&quot;data&quot;:{
+        ///         &quot;data&quot;:&quot;success&quot;,
+        ///         &quot;requestId&quot;:&quot;****&quot;
+        ///     }</para>
+        /// </summary>
+        [NameInMap("data")]
+        [Validation(Required=false)]
+        public string Data { get; set; }
+
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>OBJECT_NOT_EXIST</para>
+        /// </summary>
+        [NameInMap("errorCode")]
+        [Validation(Required=false)]
+        public string ErrorCode { get; set; }
+
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>this project is not exist.</para>
+        /// </summary>
+        [NameInMap("errorMsg")]
+        [Validation(Required=false)]
+        public string ErrorMsg { get; set; }
+
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>200</para>
+        /// </summary>
+        [NameInMap("httpCode")]
+        [Validation(Required=false)]
+        public int? HttpCode { get; set; }
+
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>73207140-0FD5-588A-B11A-3CE093924196</para>
+        /// </summary>
+        [NameInMap("requestId")]
+        [Validation(Required=false)]
+        public string RequestId { get; set; }
+
+    }
+
+}